requirements.txt
google-api-python-client
openai
python-dateutil
python-dotenv
click
